% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missing.R
\name{count_all_missing_by_group}
\alias{count_all_missing_by_group}
\title{Count Missing Values by Group}
\usage{
count_all_missing_by_group(data, group_col, .groups = "drop")
}
\arguments{
\item{data}{A data frame or tibble containing the data.}

\item{group_col}{The column within \code{data} to group by. Specify unquoted column name.}

\item{.groups}{(Optional) Controls whether the output is grouped by \code{group_col}. Defaults to \code{"drop"}.}
}
\value{
A tibble with each unique level of \code{group_col} and a count of \code{NA}s for each column in \code{data}, excluding the grouping column.
}
\description{
Computes the number of missing values in each column for each level of a specified grouping column.
}
\details{
This function groups data by the specified \code{group_col}, then counts the missing values in each column within each group.
It uses \code{dplyr} functions \code{group_by()} and \code{summarize()} for efficient grouping and summarization.
}
\examples{
# Example with the airquality dataset
count_all_missing_by_group(airquality, Month)

# Piping example
airquality |> count_all_missing_by_group(Month)

# Keeping grouped output
count_all_missing_by_group(airquality, Month, .groups = "keep")

}
